Rename services in ClientApp to follow convention
