+++ src/SIL.XForge.Scripture/ClientApp/src/app/app.module.ts
@@ -2,6 +2,7 @@ import { OverlayContainer } from '@angular/cdk/overlay';
 import { DatePipe } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ErrorHandler, NgModule } from '@angular/core';
+import { MatRipple } from '@angular/material/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
@@ -12,14 +13,13 @@ import { translocoMarkupRouterLinkRenderer } from 'ngx-transloco-markup-router-l
 import { AvatarComponent } from 'xforge-common/avatar/avatar.component';
 import { EditNameDialogComponent } from 'xforge-common/edit-name-dialog/edit-name-dialog.component';
 import { ErrorDialogComponent } from 'xforge-common/error-dialog/error-dialog.component';
-import { ExceptionHandlingService } from 'xforge-common/exception-handling-service';
+import { ExceptionHandlingService } from 'xforge-common/exception-handling.service';
 import { FeatureFlagsDialogComponent } from 'xforge-common/feature-flags/feature-flags-dialog.component';
 import { EmTextTranspiler } from 'xforge-common/i18n-transpilers/em-text.transpiler';
 import { InAppRootOverlayContainer } from 'xforge-common/overlay-container';
 import { SupportedBrowsersDialogComponent } from 'xforge-common/supported-browsers-dialog/supported-browsers-dialog.component';
 import { UICommonModule } from 'xforge-common/ui-common.module';
 import { XForgeCommonModule } from 'xforge-common/xforge-common.module';
-import { MatRipple } from '@angular/material/core';
 import { environment } from '../environments/environment';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -27,6 +27,7 @@ import { CheckingModule } from './checking/checking.module';
 import { ConnectProjectComponent } from './connect-project/connect-project.component';
 import { CoreModule } from './core/core.module';
 import { JoinComponent } from './join/join.component';
+import { MyProjectsComponent } from './my-projects/my-projects.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { ProjectComponent } from './project/project.component';
 import { ScriptureChooserDialogComponent } from './scripture-chooser-dialog/scripture-chooser-dialog.component';
@@ -34,7 +35,6 @@ import { DeleteProjectDialogComponent } from './settings/delete-project-dialog/d
 import { SettingsComponent } from './settings/settings.component';
 import { SharedModule } from './shared/shared.module';
 import { TextNoteDialogComponent } from './shared/text/text-note-dialog/text-note-dialog.component';
-import { MyProjectsComponent } from './my-projects/my-projects.component';
 import { SyncComponent } from './sync/sync.component';
 import { TranslateModule } from './translate/translate.module';
 import { UsersModule } from './users/users.module';
