+++ src/SIL.XForge.Scripture/ClientApp/src/app/shared/book-multi-select/book-multi-select.component.spec.ts
@@ -3,8 +3,8 @@ import { MatChipsModule } from '@angular/material/chips';
 import { ActivatedRoute } from '@angular/router';
 import { of } from 'rxjs';
 import { mock, when } from 'ts-mockito';
-import { configureTestingModule, TestTranslocoModule } from 'xforge-common/test-utils';
-import { ProgressService, TextProgress } from '../progress-service/progress-service';
+import { TestTranslocoModule, configureTestingModule } from 'xforge-common/test-utils';
+import { ProgressService, TextProgress } from '../progress-service/progress.service';
 import { BookMultiSelectComponent } from './book-multi-select.component';
 
 const mockedActivatedRoute = mock(ActivatedRoute);
