+++ src/SIL.XForge.Scripture/ClientApp/src/app/shared/book-multi-select/book-multi-select.component.ts
@@ -6,7 +6,7 @@ import { Canon } from '@sillsdev/scripture';
 import { filter, firstValueFrom, map } from 'rxjs';
 import { SubscriptionDisposable } from 'xforge-common/subscription-disposable';
 import { UICommonModule } from 'xforge-common/ui-common.module';
-import { ProgressService } from '../progress-service/progress-service';
+import { ProgressService } from '../progress-service/progress.service';
 
 export interface BookOption {
   bookNum: number;
