+++ src/SIL.XForge.Scripture/ClientApp/src/main.ts
@@ -1,7 +1,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { ExceptionHandlingService } from 'xforge-common/exception-handling-service';
+import { ExceptionHandlingService } from 'xforge-common/exception-handling.service';
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
