+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/exception-handling.service.ts
